apps/backend/workspace-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
PyJWT[crypto]>=2.8.0
